Extract `getSochainUrl`
