+++ src/main/api/ledger/bitcoin/transaction.ts
@@ -7,8 +7,9 @@ import { BaseAmount } from '@xchainjs/xchain-util'
 import * as Bitcoin from 'bitcoinjs-lib'
 import * as E from 'fp-ts/lib/Either'
 
+import { getSochainUrl } from '../../../../shared/api/sochain'
 import { LedgerError, LedgerErrorId, Network } from '../../../../shared/api/types'
-import { getHaskoinApiUrl, getSochainUrl } from '../../../../shared/bitcoin/client'
+import { getHaskoinApiUrl } from '../../../../shared/bitcoin/client'
 import { toClientNetwork } from '../../../../shared/utils/client'
 import { isError } from '../../../../shared/utils/guard'
 import { getDerivationPath } from './common'
