+++ src/renderer/services/bitcoin/common.ts
@@ -7,7 +7,8 @@ import * as Rx from 'rxjs'
 import { Observable } from 'rxjs'
 import * as RxOp from 'rxjs/operators'
 
-import { getHaskoinApiUrl, getSochainUrl } from '../../../shared/bitcoin/client'
+import { getSochainUrl } from '../../../../shared/api/sochain'
+import { getHaskoinApiUrl } from '../../../shared/bitcoin/client'
 import { isError } from '../../../shared/utils/guard'
 import { clientNetwork$ } from '../app/service'
 import * as C from '../clients'
